+++ src/calendarGUI.hpp
@@ -25,22 +25,22 @@
 #define EVENTEDITOR_RETURN_CONFIRM 1
 #define EVENTDELETE_RETURN_EXIT 0
 #define EVENTDELETE_RETURN_CONFIRM 1
-void viewCalendar(int dateselection=0);
-int viewMonthCalendar(int dateselection=0);
-int viewWeekCalendar();
-int viewWeekCalendarSub(Menu* menu, int* y, int* m, int* d, int* jumpToSel, int* keepMenuSel);
+void viewCalendar(short dateselection=0);
+short viewMonthCalendar(short dateselection=0);
+short viewWeekCalendar();
+short viewWeekCalendarSub(Menu* menu, short* y, short* m, short* d, short* jumpToSel, short* keepMenuSel);
 void viewEvents(int y, int m, int d);
 void viewEvent(CalendarEvent* event, int istask=0);
 int viewEventsSub(Menu* menu, int y, int m, int d);
 void fillInputDate(int yr, int m, int d, char* buffer);
 void fillInputTime(int h, int m, int s, char* buffer);
 int eventEditor(int y, int m, int d, int type=EVENTEDITORTYPE_ADD, CalendarEvent* event=NULL, int istask=0);
-void drawCalendar(int year, int month, int d, int show_event_count, int* eventcount, int* busydays, int* bufmonth, int* bufyear);
+void drawCalendar(int year, int month, int d, int show_event_count, short* eventcount, short* busydays, short* bufmonth, short* bufyear);
 int moveEvent(CalendarEvent* events, int count, int pos, int isCopy=0);
 int deleteEventUI(int y, int m, int d, CalendarEvent* events, int count, int pos, int istask=0);
 int deleteAllEventUI(int y, int m, int d, int istask=0);
-int chooseCalendarDate(int *yr, int *m, int *d, char* message, char* message2, int graphical = 0);
-void invalidFieldMsg(int istime);
+short chooseCalendarDate(short *yr, short *m, short *d, char* message, char* message2, short graphical = 0);
+void invalidFieldMsg(short istime);
 void setEventChrono(CalendarEvent* event);
 int changeEventCategory(CalendarEvent* event);
 void viewNthEventOnDay(EventDate* date, int pos);
