+++ src/calendarProvider.hpp
@@ -25,27 +25,27 @@
 
 typedef struct // Event date definition
 {
-  unsigned int day=0;
-  unsigned int month=0;
-  unsigned int year=0;
+  unsigned short day=0;
+  unsigned short month=0;
+  unsigned short year=0;
 } EventDate;
 
 typedef struct // Event time definition
 {
-  unsigned int hour=0;
-  unsigned int minute=0;
-  unsigned int second=0;
+  unsigned short hour=0;
+  unsigned short minute=0;
+  unsigned short second=0;
 } EventTime;
 
 typedef struct // Defines what a calendar event contains
 {
-  unsigned int category=0;
-  //unsigned int daterange=0; //not used for now
+  unsigned short category=0;
+  //unsigned short daterange=0; //not used for now
   EventDate startdate;
   EventDate enddate;
-  //unsigned int dayofweek=1; // not used for now
-  unsigned int repeat=0;
-  unsigned int timed=1; //full-day = 0, timed = 1
+  //unsigned short dayofweek=1; // not used for now
+  unsigned short repeat=0;
+  unsigned short timed=1; //full-day = 0, timed = 1
   EventTime starttime;
   EventTime endtime;
   unsigned char title[25]; //can't be 21, because otherwise somehow the location will replace the last chars of title
@@ -55,10 +55,10 @@ typedef struct // Defines what a calendar event contains
 
 typedef struct // a simplified calendar event, for use when the memory available is little and one needs to show many events
 {
-  unsigned int category=0;
+  unsigned short category=0;
   EventDate startdate;
   unsigned char title[25];
-  unsigned int origpos=0; //position in original file (useful for search results). zero based.
+  unsigned short origpos=0; //position in original file (useful for search results). zero based.
 } SimpleCalendarEvent;
 // end of type definitions
 
@@ -67,17 +67,17 @@ void charToCalEvent(unsigned char* src, CalendarEvent* calEvent);
 void charToSimpleCalEvent(unsigned char* src, SimpleCalendarEvent* calEvent);
 void filenameFromDate(EventDate* date, char* filename);
 void smemFilenameFromDate(EventDate* date, unsigned short* filename, const char* folder);
-int AddEvent(CalendarEvent* calEvent, const char* folder, int secondCall=0);
-int ReplaceEventFile(EventDate *startdate, CalendarEvent* newEvents, const char* folder, int count);
-void RemoveEvent(EventDate *startdate, CalendarEvent* events, const char* folder, int count, int calEventPos);
+short AddEvent(CalendarEvent* calEvent, const char* folder, short secondCall=0);
+short ReplaceEventFile(EventDate *startdate, CalendarEvent* newEvents, const char* folder, short count);
+void RemoveEvent(EventDate *startdate, CalendarEvent* events, const char* folder, short count, short calEventPos);
 void RemoveDay(EventDate* date, const char* folder);
-int GetEventsForDate(EventDate* startdate, const char* folder, CalendarEvent* calEvents, int limit=0, SimpleCalendarEvent* simpleCalEvents = NULL, int startArray=0);
-void GetEventCountsForMonth(int year, int month, int* buffer, int* busydays);
-int SearchEventsOnDay(EventDate* date, const char* folder, SimpleCalendarEvent* calEvents, char* needle, int limit);
-int SearchEventsOnMonth(int y, int m, const char* folder, SimpleCalendarEvent* calEvents, char* needle, int limit);
-int SearchEventsOnYearOrMonth(int y, int m, const char* folder, SimpleCalendarEvent* calEvents, char* needle, int limit, int arraystart=0);
+short GetEventsForDate(EventDate* startdate, const char* folder, CalendarEvent* calEvents, short limit=0, SimpleCalendarEvent* simpleCalEvents = NULL, short startArray=0);
+void GetEventCountsForMonth(short year, short month, short* buffer, short* busydays);
+short SearchEventsOnDay(EventDate* date, const char* folder, SimpleCalendarEvent* calEvents, char* needle, short limit);
+short SearchEventsOnMonth(short y, short m, const char* folder, SimpleCalendarEvent* calEvents, char* needle, short limit);
+short SearchEventsOnYearOrMonth(short y, short m, const char* folder, SimpleCalendarEvent* calEvents, char* needle, short limit, short arraystart=0);
 void repairEventsFile(char* name, const char* folder, int* checkedevents, int* problemsfound);
-void setDBneedsRepairFlag(int value);
-int getDBneedsRepairFlag();
+void setDBneedsRepairFlag(short value);
+short getDBneedsRepairFlag();
 
 #endif
